Fix errors in testcases
