+++ zkmemory/src/commitment/mod.rs
@@ -4,9 +4,5 @@ pub mod extends;
 pub mod kzg;
 /// Commitment scheme using Merkle Tree
 pub mod merkle_tree;
-///
-pub mod poseidon;
-///
-pub mod poseidon_constants;
 /// Commitment scheme using Verkle Tree
 pub mod verkle_tree;
